separated parse and scrape into files
